refactor(install-dynamic-plugins): adopt cli-module convention while keeping the bundled bin
Aligns with the @backstage/cli-module-* convention without losing the
self-contained bundled artifact:

- Rename to @red-hat-developer-hub/cli-module-install-dynamic-plugins,
  backstage.role: cli-module.
- src/installer.ts holds the install pipeline and main() (formerly
  src/index.ts).
- New src/index.ts default-exports `createCliModule(...)` registering an
  `install` command whose loader is src/command.ts. Exposes the package
  through backstage-cli discovery — `backstage-cli install <dir>` works
  when the package is a dependency.
- src/cli.ts (the esbuild entry) keeps invoking installer.main() directly,
  so the bundled .cjs stays self-contained: no @backstage/cli-node and no
  keytar gymnastics in the bin path.
- Build is dual now — `backstage-cli package build && node esbuild.config.mjs`.
  backstage-cli emits dist/index.cjs.js (the cli-module export) and the
  unbundled supporting modules; esbuild emits dist/install-dynamic-plugins.cjs
  (the standalone bin). Both are published.
- bin shim's installed branch now requires the bundled .cjs explicitly
  rather than going through `main` — that keeps direct/npx/init-container
  invocations at ~60 ms cold start instead of paying the cli-module
  dispatch cost.
- main() now takes optional `args` and `programName` so the cli-module
  loader can pass the command's argv slice and have `--help` print the
  real invocation (`install-dynamic-plugins install …`).
- @backstage/cli-node added as a runtime dependency. It is only loaded by
  the cli-module discovery path; the bundled bin never imports it.

166/166 tests pass; tsc/lint/prettier/api-reports clean. Bundle size
unchanged at ~267 KB.
