+++ workspaces/install-dynamic-plugins/.changeset/initial-release.md
@@ -1,5 +1,5 @@
 ---
-'@red-hat-developer-hub/install-dynamic-plugins': minor
+'@red-hat-developer-hub/cli-module-install-dynamic-plugins': minor
 ---
 
-Initial release. TypeScript/Node.js port of the RHDH init-container installer (originally Python; see [redhat-developer/rhdh#4574](https://github.com/redhat-developer/rhdh/pull/4574)). Ships as a single bundled `.cjs` with `tar` and `yaml` baked in; relies on `skopeo` and `npm` at runtime. CLI surface, env vars, on-disk layout, `plugin-hash` format, and tar/OCI security guards are byte-compatible with the previous Python implementation, so existing RHDH installs upgrade in place without forcing reinstalls.
+Initial release. TypeScript/Node.js port of the RHDH init-container installer (originally Python; see [redhat-developer/rhdh#4574](https://github.com/redhat-developer/rhdh/pull/4574)), packaged as a Backstage CLI module. The `install` command is registered through `createCliModule` so the package is auto-discovered by `backstage-cli` when listed as a dependency. The package also ships a self-contained esbuild bundle as its `bin`, so direct `npx install-dynamic-plugins <dir>` invocations (and RHDH's init-container `COPY` of the `.cjs`) stay fast and don't require `@backstage/cli-node` at runtime. Env vars, on-disk layout, `plugin-hash` format, and tar/OCI security guards are byte-compatible with the previous Python implementation.
