+++ workspaces/install-dynamic-plugins/packages/install-dynamic-plugins/cli-report.md
@@ -1,4 +1,4 @@
-## CLI Report file for "@red-hat-developer-hub/install-dynamic-plugins"
+## CLI Report file for "@red-hat-developer-hub/cli-module-install-dynamic-plugins"
 
 > Do not edit this file. It is a report generated by `yarn build:api-reports`
