+++ workspaces/install-dynamic-plugins/packages/install-dynamic-plugins/report.api.md
@@ -1,0 +1,13 @@
+## API Report File for "@red-hat-developer-hub/cli-module-install-dynamic-plugins"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { CliModule } from '@backstage/cli-node';
+
+// @public
+const _default: CliModule;
+export default _default;
+
+// (No @packageDocumentation comment for this package)
+```
