+++ workspaces/install-dynamic-plugins/packages/install-dynamic-plugins/src/finalize-install.test.ts
@@ -23,7 +23,7 @@ import {
 } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
-import { finalizeInstall } from './index';
+import { finalizeInstall } from './installer';
 import { GLOBAL_CONFIG_FILENAME } from './types';
 
 describe('finalizeInstall', () => {
