+++ workspaces/install-dynamic-plugins/packages/install-dynamic-plugins/src/installer.ts
@@ -1,0 +1,650 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { existsSync } from 'node:fs';
+import * as fs from 'node:fs/promises';
+import * as os from 'node:os';
+import * as path from 'node:path';
+import { cli } from 'cleye';
+import { parse as parseYaml, stringify as stringifyYaml } from 'yaml';
+import {
+  cleanupCatalogIndexTemp,
+  extractCatalogIndex,
+  extractExtraCatalogIndex,
+  parseExtraCatalogIndexImages,
+} from './catalog-index';
+import {
+  getNpmWorkers,
+  getWorkers,
+  mapConcurrent,
+  type Outcome,
+} from './concurrency';
+import { InstallException } from './errors';
+import { OciImageCache } from './image-cache';
+import { installNpmPlugin } from './installer-npm';
+import { installOciPlugin } from './installer-oci';
+import { createLock, registerLockCleanup, removeLock } from './lock-file';
+import { log } from './log';
+import {
+  deepMerge,
+  filterDisabledOciPlugins,
+  mergePlugin,
+  preMergeOciDisabledState,
+} from './merger';
+import { computePluginHash } from './plugin-hash';
+import { Skopeo } from './skopeo';
+import {
+  CONFIG_HASH_FILE,
+  DPDY_FILENAME,
+  type DynamicPluginsConfig,
+  effectivePullPolicy,
+  GLOBAL_CONFIG_FILENAME,
+  LOCK_FILENAME,
+  OCI_PROTO,
+  type Plugin,
+  type PluginMap,
+  type PluginSpec,
+  PullPolicy,
+} from './types';
+import { fileExists, isPlainObject } from './util';
+
+const CONFIG_FILE = 'dynamic-plugins.yaml';
+
+/**
+ * Entry point for the install command. `args` is the argv slice after the
+ * command path; `programName` overrides the `--help` usage line so it matches
+ * the actual invocation (e.g. `install-dynamic-plugins install` when called
+ * via the cli-module discovery path). Exported for the cli-module command
+ * loader and unit tests; not part of the package's public API.
+ *
+ * @internal
+ */
+export async function main(
+  args: string[] = process.argv.slice(2),
+  programName = 'install-dynamic-plugins',
+): Promise<void> {
+  const argv = cli(
+    {
+      name: programName,
+      parameters: ['<dynamic-plugins-root>'],
+      help: {
+        description:
+          'Install RHDH dynamic plugins listed in dynamic-plugins.yaml into the given directory.',
+      },
+    },
+    undefined,
+    args,
+  );
+  const root = path.resolve(argv._.dynamicPluginsRoot);
+  const lockPath = path.join(root, LOCK_FILENAME);
+  registerLockCleanup(lockPath);
+  await fs.mkdir(root, { recursive: true });
+  await createLock(lockPath);
+
+  let exitCode = 0;
+  try {
+    exitCode = await runInstaller(root);
+  } finally {
+    await cleanupCatalogIndexTemp(root).catch(() => undefined);
+    await removeLock(lockPath).catch(() => undefined);
+  }
+  process.exit(exitCode);
+}
+
+async function runInstaller(root: string): Promise<number> {
+  const skopeo = new Skopeo();
+  const workers = getWorkers();
+  log(`======= Workers: ${workers} (CPUs: ${os.cpus().length})`);
+
+  // Resolve the config file path against CWD at startup so the dependency on
+  // CWD is explicit in the operator log; includes are resolved relative to
+  // the config file's directory (matches the Python installer).
+  const configFileAbs = path.resolve(CONFIG_FILE);
+  const configDir = path.dirname(configFileAbs);
+  const globalConfigFile = path.join(root, GLOBAL_CONFIG_FILENAME);
+  log(`======= Config file: ${configFileAbs}`);
+
+  const entitiesDir =
+    process.env.CATALOG_ENTITIES_EXTRACT_DIR ??
+    path.join(os.tmpdir(), 'extensions');
+  const catalogDpdy = await maybeExtractCatalogIndex(skopeo, root, entitiesDir);
+  await maybeExtractExtraCatalogIndexes(skopeo, entitiesDir);
+  const content = await loadDynamicPluginsConfig(
+    configFileAbs,
+    globalConfigFile,
+  );
+  if (!content) return 0;
+
+  const imageCache = new OciImageCache(
+    skopeo,
+    await fs.mkdtemp(path.join(os.tmpdir(), 'rhdh-oci-cache-')),
+  );
+
+  const allPlugins = await loadAllPlugins(
+    content,
+    configFileAbs,
+    configDir,
+    catalogDpdy,
+    imageCache,
+  );
+  const installed = await readInstalledPluginHashes(root);
+  const globalConfig: Record<string, unknown> = {
+    dynamicPlugins: { rootDirectory: 'dynamic-plugins-root' },
+  };
+  const { oci, npm, skipped } = categorize(allPlugins);
+  handleSkippedLocals(skipped, globalConfig);
+
+  const skipIntegrity =
+    (process.env.SKIP_INTEGRITY_CHECK ?? '').toLowerCase() === 'true';
+  const errors: string[] = [];
+  await installOci(
+    oci,
+    root,
+    imageCache,
+    installed,
+    workers,
+    globalConfig,
+    errors,
+  );
+  await installNpm(npm, root, skipIntegrity, installed, globalConfig, errors);
+
+  return finalizeInstall(
+    errors,
+    globalConfigFile,
+    globalConfig,
+    root,
+    installed,
+  );
+}
+
+/** Optional `CATALOG_INDEX_IMAGE` extraction — returns the path to the
+ * extracted `dynamic-plugins.default.yaml`, or `null` when the env var is
+ * unset. */
+async function maybeExtractCatalogIndex(
+  skopeo: Skopeo,
+  root: string,
+  entitiesDir: string,
+): Promise<string | null> {
+  const catalogImage = process.env.CATALOG_INDEX_IMAGE ?? '';
+  if (!catalogImage) return null;
+  return extractCatalogIndex(skopeo, catalogImage, root, entitiesDir);
+}
+
+/**
+ * Optional `EXTRA_CATALOG_INDEX_IMAGES` extraction. Each entry is extracted
+ * into an isolated subdirectory under `<entitiesDir>/extra/` so multiple
+ * indexes can coexist without clobbering the primary index's
+ * `catalog-entities/`. Duplicate subdirectory names within the same env-var
+ * value emit an overwrite warning (handled by `extractExtraCatalogIndex`).
+ */
+async function maybeExtractExtraCatalogIndexes(
+  skopeo: Skopeo,
+  entitiesDir: string,
+): Promise<void> {
+  const raw = process.env.EXTRA_CATALOG_INDEX_IMAGES ?? '';
+  if (!raw) return;
+  const extraParent = path.join(entitiesDir, 'extra');
+  const seen = new Map<string, string>();
+  for (const [name, imageRef] of parseExtraCatalogIndexImages(raw)) {
+    const prev = seen.get(name) ?? null;
+    seen.set(name, imageRef);
+    await extractExtraCatalogIndex(skopeo, imageRef, name, extraParent, prev);
+  }
+}
+
+/** Read and parse `dynamic-plugins.yaml`. Writes an empty global config and
+ * returns `null` for the two short-circuit cases (file missing, file empty)
+ * so the caller can early-exit with code 0. */
+async function loadDynamicPluginsConfig(
+  configFileAbs: string,
+  globalConfigFile: string,
+): Promise<DynamicPluginsConfig | null> {
+  if (!(await fileExists(configFileAbs))) {
+    log(`No ${CONFIG_FILE} found at ${configFileAbs}. Skipping.`);
+    await fs.writeFile(globalConfigFile, '');
+    return null;
+  }
+  const rawContent = await fs.readFile(configFileAbs, 'utf8');
+  const content = parseYaml(rawContent) as DynamicPluginsConfig | null;
+  if (!content) {
+    log(`${configFileAbs} is empty. Skipping.`);
+    await fs.writeFile(globalConfigFile, '');
+    return null;
+  }
+  return content;
+}
+
+/** Resolve include paths, substitute the catalog-index placeholder, merge
+ * everything into a single `PluginMap`, and compute change-detection hashes.
+ *
+ * Two-phase to match the Python pre-merge OCI-disable pass: load every
+ * include file's plugin list into memory FIRST, compute the effectively
+ * disabled OCI registries, then filter those entries out of every list
+ * before merging. Without this pass an OCI plugin marked `disabled: true`
+ * at level 1 would still trigger a `skopeo` round-trip during the level-0
+ * merge — wasted work and a footgun in restricted-network init containers.
+ */
+async function loadAllPlugins(
+  content: DynamicPluginsConfig,
+  configFileAbs: string,
+  configDir: string,
+  catalogDpdy: string | null,
+  imageCache: OciImageCache,
+): Promise<PluginMap> {
+  const allPlugins: PluginMap = {};
+  const includes = resolveIncludes(
+    content.includes ?? [],
+    configDir,
+    catalogDpdy,
+  );
+
+  const includeLists: Array<[string, PluginSpec[]]> = [];
+  for (const inc of includes) {
+    if (!(await fileExists(inc))) {
+      log(`WARNING: include file ${inc} not found, skipping`);
+      continue;
+    }
+    log(`\n======= Including plugins from ${inc}`);
+    const parsed = parseYaml(
+      await fs.readFile(inc, 'utf8'),
+    ) as DynamicPluginsConfig | null;
+    if (parsed && !isPlainObject(parsed)) {
+      throw new InstallException(`${inc} must contain a mapping`);
+    }
+    const plugins = parsed?.plugins ?? [];
+    if (!Array.isArray(plugins)) {
+      throw new InstallException(
+        `${inc} must contain a 'plugins' list (got ${typeof plugins})`,
+      );
+    }
+    includeLists.push([inc, plugins]);
+  }
+  const mainPlugins = content.plugins ?? [];
+
+  const disabledRegistries = preMergeOciDisabledState(
+    includeLists,
+    mainPlugins,
+    configFileAbs,
+  );
+
+  for (const [inc, plugins] of includeLists) {
+    for (const plugin of filterDisabledOciPlugins(
+      plugins,
+      disabledRegistries,
+    )) {
+      await mergePlugin(plugin, allPlugins, inc, /* level */ 0, imageCache);
+    }
+  }
+  for (const plugin of filterDisabledOciPlugins(
+    mainPlugins,
+    disabledRegistries,
+  )) {
+    await mergePlugin(
+      plugin,
+      allPlugins,
+      configFileAbs,
+      /* level */ 1,
+      imageCache,
+    );
+  }
+
+  for (const p of Object.values(allPlugins)) {
+    p.plugin_hash = computePluginHash(p);
+  }
+  return allPlugins;
+}
+
+function resolveIncludes(
+  rawIncludes: readonly string[],
+  configDir: string,
+  catalogDpdy: string | null,
+): string[] {
+  const includes = rawIncludes.map(inc =>
+    path.isAbsolute(inc) ? inc : path.resolve(configDir, inc),
+  );
+  if (catalogDpdy) {
+    const idx = includes.findIndex(inc => path.basename(inc) === DPDY_FILENAME);
+    if (idx !== -1) includes[idx] = catalogDpdy;
+  }
+  return includes;
+}
+
+/**
+ * Write the global config, prune obsolete plugin dirs, and compute the exit
+ * code. Exported for unit tests; not part of the package's public API.
+ *
+ * @internal
+ */
+export async function finalizeInstall(
+  errors: string[],
+  globalConfigFile: string,
+  globalConfig: Record<string, unknown>,
+  root: string,
+  installed: Map<string, string>,
+): Promise<number> {
+  if (errors.length > 0) {
+    log(`\n======= ${errors.length} plugin(s) failed:`);
+    for (const err of errors) log(`  - ${err}`);
+    // Skip writing the global config and cleaning up previously-installed
+    // plugin dirs so the filesystem does not end up in an "almost valid"
+    // state. Exit 1 is enough for init containers to block startup, but a
+    // manual restart of the main container (or a deployment that does not
+    // enforce init-container success) could otherwise pick up a partial
+    // config — e.g. a frontend plugin without its backend dep, yielding a
+    // broken UI. Preserving the prior state makes the next run a clean retry.
+    log(
+      `\n======= Skipping ${GLOBAL_CONFIG_FILENAME} write and cleanup because of install failures. ` +
+        `Fix the errors above and re-run; the previous successful state is preserved.`,
+    );
+    return 1;
+  }
+
+  await fs.writeFile(globalConfigFile, stringifyYaml(globalConfig));
+  await cleanupRemoved(root, installed);
+
+  log('\n======= All plugins installed successfully');
+  return 0;
+}
+
+type Categorized = {
+  oci: Plugin[];
+  npm: Plugin[];
+  skipped: Plugin[];
+};
+
+function categorize(allPlugins: PluginMap): Categorized {
+  const oci: Plugin[] = [];
+  const npm: Plugin[] = [];
+  const skipped: Plugin[] = [];
+  for (const plugin of Object.values(allPlugins)) {
+    if (plugin.disabled) {
+      log(`\n======= Skipping disabled plugin ${plugin.package}`);
+      continue;
+    }
+    if (plugin.package.startsWith(OCI_PROTO)) {
+      oci.push(plugin);
+      continue;
+    }
+    if (plugin.package.startsWith('./')) {
+      const localPath = path.join(process.cwd(), plugin.package.slice(2));
+      if (existsSync(localPath)) npm.push(plugin);
+      else skipped.push(plugin);
+      continue;
+    }
+    npm.push(plugin);
+  }
+  return { oci, npm, skipped };
+}
+
+function handleSkippedLocals(
+  skipped: Plugin[],
+  globalConfig: Record<string, unknown>,
+): void {
+  if (skipped.length === 0) return;
+  log(
+    `\n======= Skipping ${skipped.length} local plugins (directories not found)`,
+  );
+  for (const plugin of skipped) {
+    const abs = path.join(process.cwd(), plugin.package.slice(2));
+    log(`\t==> ${plugin.package} (not found at ${abs})`);
+    if (isPlainObject(plugin.pluginConfig)) {
+      deepMerge(plugin.pluginConfig, globalConfig);
+    }
+  }
+}
+
+type InstallOutcome = {
+  pluginPath: string | null;
+  pluginConfig: Record<string, unknown>;
+};
+
+async function installOci(
+  plugins: Plugin[],
+  root: string,
+  imageCache: OciImageCache,
+  installed: Map<string, string>,
+  workers: number,
+  globalConfig: Record<string, unknown>,
+  errors: string[],
+): Promise<void> {
+  await runInstallPipeline({
+    plugins,
+    workers,
+    label: 'OCI',
+    installFn: plugin => installOciPlugin(plugin, root, imageCache, installed),
+    installed,
+    globalConfig,
+    errors,
+  });
+}
+
+async function installNpm(
+  plugins: Plugin[],
+  root: string,
+  skipIntegrity: boolean,
+  installed: Map<string, string>,
+  globalConfig: Record<string, unknown>,
+  errors: string[],
+): Promise<void> {
+  // `npm pack` writes the tarball to `cwd` with a package-derived filename
+  // (`<name>-<version>.tgz`), so concurrent invocations against different
+  // packages don't clash on the filename. The npm CLI cache
+  // (`~/.npm/_cacache`) handles its own locking. Cap defaults to 3 to keep
+  // the registry happy — override with `DYNAMIC_PLUGINS_NPM_WORKERS=1` to
+  // restore the original sequential behaviour.
+  await runInstallPipeline({
+    plugins,
+    workers: getNpmWorkers(),
+    label: 'NPM',
+    installFn: plugin =>
+      installNpmPlugin(plugin, root, skipIntegrity, installed),
+    installed,
+    globalConfig,
+    errors,
+  });
+}
+
+type RunInstallPipelineArgs = {
+  plugins: Plugin[];
+  workers: number;
+  label: 'OCI' | 'NPM';
+  installFn: (plugin: Plugin) => Promise<InstallOutcome>;
+  installed: Map<string, string>;
+  globalConfig: Record<string, unknown>;
+  errors: string[];
+};
+
+/**
+ * Shared install pipeline used by both `installOci` and `installNpm`:
+ *   1. Synchronous pre-pass that short-circuits "definitely no-op" plugins
+ *      (hash present, no force, pull policy not Always) without spinning
+ *      up the worker pool — avoids the parallel-skopeo / parallel-npm-pack
+ *      overhead in the steady-state restart case.
+ *   2. `mapConcurrent` over the plugins that actually need work, capped by
+ *      `workers`.
+ *   3. Single-pass over the outcomes that records errors and merges plugin
+ *      configs into the global config.
+ *
+ * Keeping both categories on this shared body so a behaviour change (a new
+ * fast-path filter, a different log format, an extra error pathway) does
+ * not have to be made twice in two slightly-divergent copies.
+ */
+async function runInstallPipeline(args: RunInstallPipelineArgs): Promise<void> {
+  const {
+    plugins,
+    workers,
+    label,
+    installFn,
+    installed,
+    globalConfig,
+    errors,
+  } = args;
+  if (plugins.length === 0) return;
+
+  const needsWork = partitionDefinitelyNoOp(
+    plugins,
+    installed,
+    globalConfig,
+    errors,
+  );
+  if (needsWork.length === 0) return;
+
+  const workerSuffix = workers === 1 ? '' : 's';
+  log(
+    `\n======= Installing ${needsWork.length} ${label} plugin(s) (${workers} worker${workerSuffix})`,
+  );
+
+  const results = await mapConcurrent(needsWork, workers, async plugin => {
+    log(`\n======= Installing ${label} plugin ${plugin.package}`);
+    return installFn(plugin);
+  });
+
+  applyInstallOutcomes(results, globalConfig, errors);
+}
+
+/**
+ * Synchronous pre-pass: drop plugins that are definitely a no-op (hash on
+ * disk, not forced, not Always-pull) without paying the worker-pool /
+ * Promise overhead, and return the remaining plugins that actually need
+ * installation work. Side-effect: removes the no-op plugins from
+ * `installed` and merges their `pluginConfig` into `globalConfig`.
+ */
+function partitionDefinitelyNoOp(
+  plugins: Plugin[],
+  installed: Map<string, string>,
+  globalConfig: Record<string, unknown>,
+  errors: string[],
+): Plugin[] {
+  const needsWork: Plugin[] = [];
+  for (const plugin of plugins) {
+    if (definitelyNoOp(plugin, installed)) {
+      log(`\t==> ${plugin.package}: already installed, skipping`);
+      installed.delete(plugin.plugin_hash);
+      mergeConfigSafely(
+        plugin.pluginConfig,
+        globalConfig,
+        plugin.package,
+        errors,
+      );
+    } else {
+      needsWork.push(plugin);
+    }
+  }
+  return needsWork;
+}
+
+/**
+ * Drain a `mapConcurrent` outcome list: record errors, merge configs into
+ * the global config, log success lines. Pulled out of `runInstallPipeline`
+ * to keep that orchestrator small enough to read top-to-bottom.
+ */
+function applyInstallOutcomes(
+  results: ReadonlyArray<Outcome<InstallOutcome, Plugin>>,
+  globalConfig: Record<string, unknown>,
+  errors: string[],
+): void {
+  for (const outcome of results) {
+    if (!outcome.ok) {
+      errors.push(`${outcome.item.package}: ${outcome.error.message}`);
+      log(`\t==> ERROR: ${outcome.item.package}: ${outcome.error.message}`);
+      continue;
+    }
+    const { value, item } = outcome;
+    const merged = mergeConfigSafely(
+      value.pluginConfig,
+      globalConfig,
+      item.package,
+      errors,
+    );
+    if (merged && value.pluginPath) {
+      log(`\t==> Installed ${item.package}`);
+    }
+  }
+}
+
+/**
+ * Merge `pluginConfig` into `globalConfig` if it is a plain object. Returns
+ * `false` and pushes onto `errors` when `deepMerge` raises a key conflict —
+ * the caller uses this to skip the "Installed" success log so the operator
+ * sees only the error line for the affected plugin.
+ */
+function mergeConfigSafely(
+  pluginConfig: Record<string, unknown> | undefined,
+  globalConfig: Record<string, unknown>,
+  pkg: string,
+  errors: string[],
+): boolean {
+  if (!isPlainObject(pluginConfig)) return true;
+  try {
+    deepMerge(pluginConfig, globalConfig);
+    return true;
+  } catch (err) {
+    errors.push(`${pkg}: ${(err as Error).message}`);
+    return false;
+  }
+}
+
+/**
+ * Cheap synchronous check: a plugin is "definitely" a no-op when its hash
+ * is already on disk, the user did not force a re-download, and the pull
+ * policy doesn't demand a remote-digest comparison. ALWAYS-pull plugins
+ * still go through the regular install path because they need a
+ * `skopeo inspect` to compare local vs remote digest.
+ *
+ * Type guard: narrows `plugin.plugin_hash` to a non-undefined `string`
+ * inside the `if (definitelyNoOp(...))` branch so the caller does not
+ * need a `as string` cast on the subsequent `installed.delete` call.
+ */
+function definitelyNoOp(
+  plugin: Plugin,
+  installed: Map<string, string>,
+): plugin is Plugin & { plugin_hash: string } {
+  if (!plugin.plugin_hash || !installed.has(plugin.plugin_hash)) return false;
+  if (plugin.forceDownload) return false;
+  return effectivePullPolicy(plugin) !== PullPolicy.ALWAYS;
+}
+
+async function cleanupRemoved(
+  root: string,
+  installed: Map<string, string>,
+): Promise<void> {
+  for (const [, dir] of installed) {
+    const pluginDir = path.join(root, dir);
+    log(`\n======= Removing obsolete plugin ${dir}`);
+    await fs.rm(pluginDir, { recursive: true, force: true });
+  }
+}
+
+async function readInstalledPluginHashes(
+  root: string,
+): Promise<Map<string, string>> {
+  const installed = new Map<string, string>();
+  let entries: string[];
+  try {
+    entries = await fs.readdir(root);
+  } catch {
+    return installed;
+  }
+  for (const entry of entries) {
+    const hashFile = path.join(root, entry, CONFIG_HASH_FILE);
+    try {
+      const hash = (await fs.readFile(hashFile, 'utf8')).trim();
+      if (hash) installed.set(hash, entry);
+    } catch {
+      /* not a plugin dir */
+    }
+  }
+  return installed;
+}
